Fix README to be proper RST syntax
The README file was in RST, but had a formatting issue with the title.
The usage of the README is setup was incorrectly setting the content
type to markdown, rather than RST. This seems to have caused our twine
check to validate the package as fine, but PyPi apparently checks
package metadata a different way and rejected the package upload due to
issues with the README file.

This corrects the RST syntax, renames the file from txt to rst to
accurately reflect its content type, and updates the settings up
setup.py to correctly report the content as RST.

Change-Id: I5135a709410372aa938446563e8eca70b5eea985
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>